Move one more function to sm_sir
